# perf: Speed up sync index database