`reader.iter_from` spend too much time to scan the database (scan every 1 block), now only scan the database when need (when commit batch of blocks, which about 200 blocks). It will increase sync speed about 14 times.

Also `Fix --wait-for-sync`

```
ckb-cli rpc get_tip_block_number
108601

# Old version
time ./target/release/ckb-cli --wait-for-sync wallet top-capacity
./target/release/ckb-cli --wait-for-sync wallet top-capacity  1792.77s user 98.92s system 97% cpu 32:27.96 total

# New version
time ckb-cli --wait-for-sync wallet top-capacity
ckb-cli --wait-for-sync wallet top-capacity  129.03s user 18.65s system 73% cpu 3:21.45 total
```