@TheWaWaR 2019-12-31
> Is the benchmark result reversed?

Oops, it is reversed.